requirements.txt
django==4.2  
celery==5.3.6  
redis==4.6.0  
psycopg==3.1.10
django-environ==0.11.2
djangorestframework==3.14.0
pytest==8.1.1
pytest-django==4.7.0
pytest-celery